# app/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/recipes": {"post": {"tags": ["recipes"], "summary": "Create Recipe", "description": "Create a new recipe.\n- **recipe**: Recipe object to be created.", "operationId": "create_recipe_recipes_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["recipes"], "summary": "Get All Recipes", "description": "Retrieve all recipes with pagination.\n- **skip**: The number of records to skip.\n- **limit**: The maximum number of records to retrieve.", "operationId": "get_all_recipes_recipes_get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "description": "Number of records to skip", "default": 0, "title": "Skip"}, "description": "Number of records to skip"}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Number of records to retrieve", "default": 10, "title": "Limit"}, "description": "Number of records to retrieve"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/recipes/name/{name}": {"get": {"tags": ["recipes"], "summary": "Get Recipe By Name", "description": "Retrieve a recipe by its name.\n- **name**: The name of the recipe.", "operationId": "get_recipe_by_name_recipes_name__name__get", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["recipes"], "summary": "Update Recipe By Name", "description": "Update a recipe by its name.\n- **name**: The name of the recipe to update.\n- **recipe**: Recipe object containing the updated data.", "operationId": "update_recipe_by_name_recipes_name__name__put", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["recipes"], "summary": "Delete Recipe By Name", "description": "Delete a recipe by its name.\n- **name**: The name of the recipe to delete.", "operationId": "delete_recipe_by_name_recipes_name__name__delete", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/recipes/id/{recipe_id}": {"get": {"tags": ["recipes"], "summary": "Get Recipe By Id", "description": "Retrieve a recipe by its ID.\n- **recipe_id**: The ID of the recipe.", "operationId": "get_recipe_by_id_recipes_id__recipe_id__get", "parameters": [{"name": "recipe_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Recipe Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["recipes"], "summary": "Update Recipe By Id", "description": "Update a recipe by its ID.\n- **recipe_id**: The ID of the recipe to update.\n- **recipe**: Recipe object containing the updated data.", "operationId": "update_recipe_by_id_recipes_id__recipe_id__put", "parameters": [{"name": "recipe_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Recipe Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipe"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["recipes"], "summary": "Delete Recipe By Id", "description": "Delete a recipe by its ID.\n- **recipe_id**: The ID of the recipe to delete.", "operationId": "delete_recipe_by_id_recipes_id__recipe_id__delete", "parameters": [{"name": "recipe_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Recipe Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Ingredient": {"properties": {"ingredient_name": {"type": "string", "title": "Ingredient Name"}, "quantity": {"type": "string", "title": "Quantity"}}, "type": "object", "required": ["ingredient_name", "quantity"], "title": "Ingredient"}, "PaginatedResponse": {"properties": {"items": {"items": {}, "type": "array", "title": "Items"}, "links": {"type": "object", "title": "Links"}, "correlationId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlationid"}}, "type": "object", "required": ["items", "links"], "title": "PaginatedResponse"}, "Recipe": {"properties": {"recipe_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Recipe Id"}, "name": {"type": "string", "title": "Name"}, "ingredients": {"items": {"$ref": "#/components/schemas/Ingredient"}, "type": "array", "title": "Ingredients"}, "steps": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Steps"}, "time_to_cook": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Time To Cook"}, "meal_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Meal Type"}, "calories": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Calories"}, "rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Rating"}, "links": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Links"}, "correlationId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlationid"}}, "type": "object", "required": ["name", "ingredients"], "title": "Recipe", "example": {"calories": 300, "ingredients": [{"ingredient_name": "Avocado", "quantity": "1 large"}, {"ingredient_name": "Bread", "quantity": "2 slices"}, {"ingredient_name": "Lime", "quantity": "1/2"}, {"ingredient_name": "Olive oil", "quantity": "1 tbsp"}, {"ingredient_name": "Salt", "quantity": "1/4 tsp"}], "links": {"delete": {"href": "/recipes/171", "method": "DELETE"}, "self": {"href": "/recipes/171"}, "update": {"href": "/recipes/171", "method": "PUT"}}, "meal_type": "breakfast", "name": "Avocado Toast", "rating": 4.4, "recipe_id": 171, "steps": "1. Toast bread. 2. Mash avocado with lime. 3. Spread on toast and season.", "time_to_cook": 10}}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}